chore: consolidate volume control into a single folder and split into components
