--- frontend/src/container/MetricsExplorer/MetricDetails/VolumeControl/VolumeControlConfig.module.scss
@@ -1,172 +1,0 @@
-.body {
-	display: flex;
-	flex-direction: column;
-	gap: 20px;
-	padding: 4px 2px;
-}
-
-.title {
-	display: flex;
-	align-items: center;
-	gap: 10px;
-}
-
-.admin-tag {
-	font-size: 10px;
-	font-weight: 600;
-	letter-spacing: 0.03em;
-	color: var(--bg-amber-400, #ffd778);
-	border: 1px solid var(--bg-amber-500, #ffcc56);
-	border-radius: 99px;
-	padding: 1px 8px;
-}
-
-/* mode cards */
-.mode-cards {
-	display: grid;
-	grid-template-columns: repeat(3, 1fr);
-	gap: 10px;
-}
-
-.mode-card {
-	display: flex;
-	flex-direction: column;
-	gap: 4px;
-	text-align: left;
-	border: 1px solid var(--bg-slate-400, #1d212d);
-	border-radius: 6px;
-	background: var(--bg-ink-300, #16181d);
-	padding: 12px;
-	cursor: pointer;
-	transition:
-		border-color 0.12s ease,
-		background 0.12s ease;
-}
-
-.mode-card:hover {
-	border-color: var(--bg-slate-200, #2c3140);
-}
-
-.mode-card-active {
-	border-color: var(--bg-robin-500, #4e74f8);
-	background: rgba(78, 116, 248, 0.08);
-}
-
-.mode-title {
-	font-size: 12.5px;
-	font-weight: 600;
-	color: var(--bg-vanilla-100, #fff);
-}
-
-.mode-desc {
-	font-size: 11px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-	line-height: 1.45;
-}
-
-/* label selector */
-.field {
-	display: flex;
-	flex-direction: column;
-	gap: 6px;
-}
-
-.field-label {
-	font-size: 12.5px;
-	font-weight: 600;
-	color: var(--bg-vanilla-100, #fff);
-}
-
-.field-hint {
-	font-size: 11px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-}
-
-.label-select {
-	width: 100%;
-	margin-top: 4px;
-}
-
-/* impact panel */
-.impact {
-	border: 1px solid var(--bg-slate-400, #1d212d);
-	border-radius: 6px;
-	background: var(--bg-ink-300, #16181d);
-	padding: 14px;
-}
-
-.impact-note {
-	font-size: 12px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-}
-
-.meters {
-	display: grid;
-	grid-template-columns: repeat(3, 1fr);
-	gap: 12px;
-}
-
-.meter {
-	display: flex;
-	flex-direction: column;
-	gap: 5px;
-}
-
-.meter-label {
-	font-size: 10px;
-	text-transform: uppercase;
-	letter-spacing: 0.04em;
-	color: var(--bg-vanilla-400, #c0c1c3);
-}
-
-.meter-value {
-	font-family: 'Geist Mono', monospace;
-	font-size: 18px;
-	color: var(--bg-vanilla-100, #fff);
-}
-
-.meter-value-good {
-	color: var(--bg-forest-400, #50e7a7);
-}
-
-/* related-asset warning */
-.warning {
-	display: flex;
-	gap: 10px;
-	padding: 11px 13px;
-	border-radius: 6px;
-	background: rgba(255, 204, 86, 0.07);
-	border: 1px solid rgba(255, 204, 86, 0.3);
-	color: var(--bg-amber-400, #ffd778);
-}
-
-.warning-title {
-	font-size: 12px;
-	font-weight: 600;
-	color: var(--bg-vanilla-100, #fff);
-}
-
-.warning-detail {
-	font-size: 11.5px;
-	color: var(--bg-vanilla-300, #e9e9e9);
-	margin-top: 4px;
-}
-
-.asset-list {
-	margin: 6px 0 0;
-	padding-left: 16px;
-	font-size: 11.5px;
-	color: var(--bg-vanilla-300, #e9e9e9);
-}
-
-/* footer */
-.footer {
-	display: flex;
-	align-items: center;
-	gap: 10px;
-	width: 100%;
-}
-
-.footer-spacer {
-	flex: 1;
-}
