--- frontend/src/container/MetricsExplorer/MetricDetails/VolumeControl/VolumeControlSection.module.scss
@@ -1,114 +1,0 @@
-.section {
-	margin-top: 16px;
-}
-
-.header {
-	display: flex;
-	align-items: center;
-	gap: 8px;
-	margin-bottom: 10px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-}
-
-.title {
-	font-size: 11px;
-	font-weight: 600;
-	text-transform: uppercase;
-	letter-spacing: 0.05em;
-}
-
-.card {
-	border: 1px solid var(--bg-slate-400, #1d212d);
-	border-radius: 6px;
-	padding: 12px 14px;
-	background: var(--bg-ink-300, #16181d);
-}
-
-.card-row {
-	display: flex;
-	align-items: center;
-	gap: 8px;
-	margin-bottom: 8px;
-}
-
-.status-dot {
-	width: 7px;
-	height: 7px;
-	border-radius: 50%;
-	flex: 0 0 auto;
-}
-
-.status-active {
-	background: var(--bg-forest-500, #25e192);
-}
-
-.status-pending {
-	background: var(--bg-amber-500, #ffcc56);
-}
-
-.card-title {
-	font-weight: 600;
-}
-
-.mode {
-	display: block;
-	font-size: 12px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-	margin-bottom: 8px;
-}
-
-.chips {
-	display: flex;
-	flex-wrap: wrap;
-	gap: 6px;
-}
-
-.chip {
-	font-family: 'Geist Mono', monospace;
-	font-size: 11px;
-	padding: 2px 8px;
-	border-radius: 4px;
-	background: var(--bg-ink-200, #23262e);
-	border: 1px solid var(--bg-slate-400, #1d212d);
-	color: var(--bg-vanilla-100, #fff);
-}
-
-.empty {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	border: 1px dashed var(--bg-slate-300, #242834);
-	border-radius: 6px;
-	padding: 14px;
-}
-
-.empty-text {
-	font-size: 12px;
-	color: var(--bg-vanilla-400, #c0c1c3);
-}
-
-.setup-button {
-	margin-top: 12px;
-}
-
-.edit-button {
-	margin-left: auto;
-}
-
-.pending-banner {
-	display: flex;
-	align-items: flex-start;
-	gap: 8px;
-	padding: 9px 12px;
-	margin-bottom: 10px;
-	border-radius: 6px;
-	background: rgba(35, 196, 248, 0.07);
-	border: 1px solid rgba(35, 196, 248, 0.25);
-	color: var(--bg-aqua-400, #4bcff9);
-}
-
-.pending-text {
-	font-size: 11.5px;
-	color: var(--bg-vanilla-300, #e9e9e9);
-	line-height: 1.45;
-}
