+++ frontend/src/container/MetricsExplorer/MetricDetails/__tests__/MetricDetails.test.tsx
@@ -78,7 +78,7 @@ jest.mock(
 );
 
 jest.mock(
-	'container/MetricsExplorer/MetricDetails/VolumeControl/VolumeControlSection',
+	'container/MetricsExplorer/VolumeControl/components/VolumeControlSection/VolumeControlSection',
 	() =>
 		function MockVolumeControlSection(): JSX.Element {
 			return <div data-testid="volume-control-section-mock">Volume Control</div>;
