+++ frontend/src/container/MetricsExplorer/VolumeControl/components/VolumeControlConfigDrawer/RelatedAssetsWarning/RelatedAssetsWarning.module.scss
@@ -1,0 +1,28 @@
+.warning {
+	display: flex;
+	gap: 10px;
+	padding: 11px 13px;
+	border-radius: 6px;
+	background: rgba(255, 204, 86, 0.07);
+	border: 1px solid rgba(255, 204, 86, 0.3);
+	color: var(--bg-amber-400, #ffd778);
+}
+
+.warning-title {
+	font-size: 12px;
+	font-weight: 600;
+	color: var(--bg-vanilla-100, #fff);
+}
+
+.warning-detail {
+	font-size: 11.5px;
+	color: var(--bg-vanilla-300, #e9e9e9);
+	margin-top: 4px;
+}
+
+.asset-list {
+	margin: 6px 0 0;
+	padding-left: 16px;
+	font-size: 11.5px;
+	color: var(--bg-vanilla-300, #e9e9e9);
+}
