+++ frontend/src/container/MetricsExplorer/VolumeControl/components/VolumeControlSection/NoRuleEmptyState/NoRuleEmptyState.module.scss
@@ -1,0 +1,17 @@
+.empty {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	border: 1px dashed var(--bg-slate-300, #242834);
+	border-radius: 6px;
+	padding: 14px;
+}
+
+.empty-text {
+	font-size: 12px;
+	color: var(--bg-vanilla-400, #c0c1c3);
+}
+
+.setup-button {
+	margin-top: 12px;
+}
