+++ frontend/src/container/MetricsExplorer/VolumeControl/components/VolumeControlSection/PendingActivationBanner/PendingActivationBanner.module.scss
@@ -1,0 +1,17 @@
+.pending-banner {
+	display: flex;
+	align-items: flex-start;
+	gap: 8px;
+	padding: 9px 12px;
+	margin-bottom: 10px;
+	border-radius: 6px;
+	background: rgba(35, 196, 248, 0.07);
+	border: 1px solid rgba(35, 196, 248, 0.25);
+	color: var(--bg-aqua-400, #4bcff9);
+}
+
+.pending-text {
+	font-size: 11.5px;
+	color: var(--bg-vanilla-300, #e9e9e9);
+	line-height: 1.45;
+}
