+++ frontend/src/container/MetricsExplorer/VolumeControl/components/VolumeControlTab/VolumeControlStats/VolumeControlStats.module.scss
@@ -1,0 +1,55 @@
+.stats {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 12px;
+}
+
+.stat {
+	display: flex;
+	flex: 1 1 0;
+	flex-direction: column;
+	gap: 6px;
+	min-width: 150px;
+	padding: 12px 16px;
+	border: 1px solid var(--bg-slate-400, #1d212d);
+	border-radius: 6px;
+	background: var(--bg-ink-400, #121317);
+}
+
+.statHero {
+	border-color: rgba(80, 231, 167, 0.4);
+	background: rgba(80, 231, 167, 0.06);
+}
+
+.statLabel {
+	font-size: 11px;
+	text-transform: uppercase;
+	letter-spacing: 0.05em;
+	color: var(--bg-vanilla-400, #c0c1c3);
+}
+
+.statValue {
+	display: flex;
+	align-items: baseline;
+	gap: 6px;
+	font-family: 'Geist Mono', monospace;
+	font-size: 22px;
+	font-weight: 600;
+	color: var(--bg-vanilla-100, #fff);
+}
+
+.statValueGood {
+	color: var(--bg-forest-400, #50e7a7);
+}
+
+.statDelta {
+	font-size: 13px;
+	font-weight: 600;
+	color: var(--bg-forest-400, #50e7a7);
+}
+
+.statUnit {
+	font-size: 13px;
+	font-weight: 500;
+	color: var(--bg-vanilla-400, #c0c1c3);
+}
