+++ frontend/src/container/MetricsExplorer/VolumeControl/components/VolumeControlTab/VolumeControlToolbar/VolumeControlToolbar.module.scss
@@ -1,0 +1,8 @@
+.toolbar {
+	display: flex;
+	justify-content: flex-end;
+}
+
+.search {
+	max-width: 320px;
+}
